+++ source/Ch08/drill/my.h
@@ -2,5 +2,5 @@ extern int foo;
 void print_foo();
 void print(int);
 void swap_v(int, int);
-void swap_r(double&, double&);
+void swap_r(int&, int&);
 void swap_cr(const int&, const int&);
